requirements.txt
flask
requests
pycryptodome
beautifulsoup4
apscheduler
